Given how many changes there are in this push I surgest that you backup your current folder structure before pulling just incase anything goes wrong.
All assets as on list have been created and placed in the scene under CDarb > Maps > CDarb combined, the scene will need to be re aranged for gameplay as I have just thrown almost everything in the scene so you know its there.

Had some issues with the key pad had to bodge the mesh in the blueprint by scaling it town 0.5 in one direction, also inserted blank keys into the key pad, was not too sure on the rest.

Updated prettymuch all your blueprints with asthetics only so there should be no reason for any of them to have stopped working but as stated I surgest you back up your full folder structure before pulling so that there is no chance of any issues.